+++ analysis/DataLoader.cc
@@ -53,7 +53,8 @@ DataLoader::DataLoader(std::string fileName,
   allBranchesOn(allBranchesOnIn),
   branchesToTurnOn(branchesToTurnOnIn),
   backwardsCompatible(backwardsCompatibleIn),
-  dataVersion(4)
+  g4dChain(nullptr),
+  dataVersion(5)
 {
   CommonCtor(fileName);
 }