+++ comparator/Compare.cc
@@ -54,6 +54,8 @@ const double        TREETOLERANCE = 0.05;
 const double OPTICSSIGMATOLERANCE = 10;
 const double   EVENTTREETOLERANCE = 1e-10;
 
+const int    EXPECTEDDATAVERSION  = 5;
+
 std::vector<Result*> Compare::Files(TFile* f1, TFile* f2)
 {
   std::vector<Result*> results; 
@@ -163,16 +165,18 @@ void Compare::Histograms(TH1* h1, TH1* h2, std::vector<Result*>& results)
 
   c->chi2   = 0.0;
   int ndof = 0; 
-  for(int i=0;i < h1->GetNbinsX(); i++)
+  for (int i=0;i < h1->GetNbinsX(); i++)
     { 
       //    std::cout << h1->GetBinContent(i) << " " << h2->GetBinContent(i) << " " << h1->GetBinError(i) << std::endl;
-      if(h1->GetBinError(i) > 0)
+      if (h1->GetBinError(i) > 0)
 	{
 	  c->chi2 += std::pow(h1->GetBinContent(i)-h2->GetBinContent(i),2)/(std::pow(h1->GetBinError(i),2)+std::pow(h2->GetBinError(i),2));
 	  ndof++;
 	}
     }
   // chi2 per dof
+  if (!std::isnormal(ndof))
+    {ndof = 1;}
   c->chi2 /= ndof;
   
   c->passed = true;
@@ -410,7 +414,7 @@ void Compare::EventTree(TTree* t1, TTree* t2, std::vector<Result*>& results,
 {
   ResultEventTree* ret = new ResultEventTree();
   ret->name            = t1->GetName();
-  ret->passed          = true; // set deafault to pass
+  ret->passed          = true; // set default to pass
   ret->objtype         = "TTree(Event)";
   ret->t1NEntries      = (int)t1->GetEntries();
   ret->t2NEntries      = (int)t2->GetEntries();
@@ -426,8 +430,8 @@ void Compare::EventTree(TTree* t1, TTree* t2, std::vector<Result*>& results,
 
   // Need to tell Event to process samplers at construction time.
   G4bool processSamplers = !samplerNames.empty();
-  Event* evtLocal1 = new Event(/*debug=*/false, processSamplers);
-  Event* evtLocal2 = new Event(/*debug=*/false, processSamplers);
+  Event* evtLocal1 = new Event(/*debug=*/false, processSamplers, EXPECTEDDATAVERSION);
+  Event* evtLocal2 = new Event(/*debug=*/false, processSamplers, EXPECTEDDATAVERSION);
   evtLocal1->SetBranchAddress(t1, &samplerNames);
   evtLocal2->SetBranchAddress(t2, &samplerNames);
 
@@ -533,6 +537,7 @@ bool Compare::Summarise(std::vector<Result*> results)
   const int titleWidth = 20;
   const int fullWidth  = titleWidth + 22;
   std::cout << std::endl;
+  std::cout << "N results: " << results.size() << std::endl;
   std::cout << "Comparison: " << std::setw(titleWidth) << "Object Name" << "   "
 	    << "Result" << std::endl;
   std::cout << std::setfill('-') << std::setw(fullWidth) << " " << std::endl;