+++ parser/optionsBase.cc
@@ -1,4 +1,4 @@
-/* 
+/*
 Beam Delivery Simulation (BDSIM) Copyright (C) Royal Holloway, 
 University of London 2001 - 2019.
 
@@ -273,6 +273,9 @@ OptionsBase::OptionsBase()
   storeTrajectoryELossSRange     = "";
   storeTrajectoryTransportationSteps = true;
   trajNoTransportation               = false;
+  storeTrajectoryLocal           = false;
+  storeTrajectoryLinks           = false;
+  storeTrajectoryIons            = false;
   
   storeSamplerAll          = false;
   storeSamplerPolarCoords  = false;